feat: adjusting tests for auto configuration
